Extract CollectionBase/Collection<T> into collection.hpp
