+++ src/realm/list.hpp
@@ -19,6 +19,8 @@
 #ifndef REALM_LIST_HPP
 #define REALM_LIST_HPP
 
+#include <realm/collection.hpp>
+
 #include <realm/obj.hpp>
 #include <realm/bplustree.hpp>
 #include <realm/obj_list.hpp>
@@ -44,200 +46,6 @@ namespace realm {
 class TableView;
 class SortDescriptor;
 class Group;
-
-// To be used in query for size. Adds nullability to size so that
-// it can be put in a NullableVector
-struct SizeOfList {
-    static constexpr size_t null_value = size_t(-1);
-
-    SizeOfList(size_t s = null_value)
-        : sz(s)
-    {
-    }
-    bool is_null()
-    {
-        return sz == null_value;
-    }
-    void set_null()
-    {
-        sz = null_value;
-    }
-    size_t size() const
-    {
-        return sz;
-    }
-    size_t sz = null_value;
-};
-
-inline std::ostream& operator<<(std::ostream& ostr, SizeOfList size_of_list)
-{
-    if (size_of_list.is_null()) {
-        ostr << "null";
-    }
-    else {
-        ostr << size_of_list.sz;
-    }
-    return ostr;
-}
-
-class CollectionBase : public ArrayParent {
-public:
-    virtual ~CollectionBase();
-    /*
-     * Operations that makes sense without knowing the specific type
-     * can be made virtual.
-     */
-    virtual size_t size() const = 0;
-    virtual bool is_null(size_t ndx) const = 0;
-    virtual Mixed get_any(size_t ndx) const = 0;
-
-    virtual Mixed min(size_t* return_ndx = nullptr) const = 0;
-    virtual Mixed max(size_t* return_ndx = nullptr) const = 0;
-    virtual Mixed sum(size_t* return_cnt = nullptr) const = 0;
-    virtual Mixed avg(size_t* return_cnt = nullptr) const = 0;
-
-    // Modifies a vector of indices so that they refer to values sorted according
-    // to the specified sort order
-    virtual void sort(std::vector<size_t>& indices, bool ascending = true) const = 0;
-    // Modifies a vector of indices so that they refer to distinct values.
-    // If 'sort_order' is supplied, the indices will refer to values in sort order,
-    // otherwise the indices will be in original order.
-    virtual void distinct(std::vector<size_t>& indices, util::Optional<bool> sort_order = util::none) const = 0;
-
-    bool is_empty() const
-    {
-        return size() == 0;
-    }
-    ObjKey get_key() const
-    {
-        return m_obj.get_key();
-    }
-    bool is_attached() const
-    {
-        return m_obj.is_valid();
-    }
-    bool has_changed() const
-    {
-        update_if_needed();
-        if (m_last_content_version != m_content_version) {
-            m_last_content_version = m_content_version;
-            return true;
-        }
-        return false;
-    }
-    void update_child_ref(size_t, ref_type new_ref) override
-    {
-        m_obj.set_int(CollectionBase::m_col_key, from_ref(new_ref));
-    }
-    ConstTableRef get_table() const
-    {
-        return m_obj.get_table();
-    }
-    ColKey get_col_key() const
-    {
-        return m_col_key;
-    }
-
-    bool operator==(const CollectionBase& other) const
-    {
-        return get_key() == other.get_key() && get_col_key() == other.get_col_key();
-    }
-
-protected:
-    template <class>
-    friend class LstIterator;
-    friend class Transaction;
-
-
-    Obj m_obj;
-    ColKey m_col_key;
-    bool m_nullable = false;
-
-    mutable std::vector<size_t> m_deleted;
-    mutable uint_fast64_t m_content_version = 0;
-    mutable uint_fast64_t m_last_content_version = 0;
-
-    CollectionBase() = default;
-    CollectionBase(bool)
-    {
-        REALM_ASSERT(false);
-    }
-    CollectionBase(const Obj& owner, ColKey col_key);
-    CollectionBase& operator=(const CollectionBase& other)
-    {
-        m_obj = other.m_obj;
-        m_col_key = other.m_col_key;
-        m_nullable = other.m_nullable;
-        return *this;
-    }
-
-    virtual bool init_from_parent() const = 0;
-
-    ref_type get_child_ref(size_t) const noexcept override;
-
-    void update_if_needed() const
-    {
-        auto content_version = m_obj.get_alloc().get_content_version();
-        if (m_obj.update_if_needed() || content_version != m_content_version) {
-            init_from_parent();
-        }
-    }
-    void update_content_version() const
-    {
-        m_content_version = m_obj.get_alloc().get_content_version();
-    }
-    // Increase index by one. I we land on and index that is deleted, keep
-    // increasing until we get to a valid entry.
-    size_t incr(size_t ndx) const
-    {
-        ndx++;
-        if (!m_deleted.empty()) {
-            auto it = m_deleted.begin();
-            auto end = m_deleted.end();
-            while (it != end && *it < ndx) {
-                ++it;
-            }
-            // If entry is deleted, increase further
-            while (it != end && *it == ndx) {
-                ++it;
-                ++ndx;
-            }
-        }
-        return ndx;
-    }
-    // Convert from virtual to real index
-    size_t adjust(size_t ndx) const
-    {
-        if (!m_deleted.empty()) {
-            // Optimized for the case where the iterator is past that last deleted entry
-            auto it = m_deleted.rbegin();
-            auto end = m_deleted.rend();
-            while (it != end && *it >= ndx) {
-                if (*it == ndx) {
-                    throw std::out_of_range("Element was deleted");
-                }
-                ++it;
-            }
-            auto diff = end - it;
-            ndx -= diff;
-        }
-        return ndx;
-    }
-    void adj_remove(size_t ndx)
-    {
-        auto it = m_deleted.begin();
-        auto end = m_deleted.end();
-        while (it != end && *it <= ndx) {
-            ++ndx;
-            ++it;
-        }
-        m_deleted.insert(it, ndx);
-    }
-    void erase_repl(Replication* repl, size_t ndx) const;
-    void move_repl(Replication* repl, size_t from, size_t to) const;
-    void swap_repl(Replication* repl, size_t ndx1, size_t ndx2) const;
-    void clear_repl(Replication* repl) const;
-};
 
 /*
  * This class implements a forward iterator over the elements in a Lst.
@@ -486,9 +294,7 @@ public:
         : CollectionBase(false)
     {
     }
-    virtual ~LstBase()
-    {
-    }
+    virtual ~LstBase() {}
     LstBasePtr clone() const
     {
         return m_obj.get_listbase_ptr(m_col_key);
@@ -953,6 +759,6 @@ inline ColumnAverageType<T> list_average(const Collection<T>& list, size_t* retu
 {
     return bptree_average(list.get_tree(), return_cnt);
 }
-}
+} // namespace realm
 
 #endif /* REALM_LIST_HPP */
